♻ refactor: 공통 dto 분리
장소, 작성자, 발신자 등 api 응답 내에서 중복적으로 사용되는 데이터를 common dto로 분리하여 재사용성 증가
