+++ src/main/java/com/ktb/cafeboo/domain/coffeechat/dto/common/SenderDto.java
@@ -1,0 +1,7 @@
+package com.ktb.cafeboo.domain.coffeechat.dto.common;
+
+public record SenderDto(
+        String userId,
+        String nickname,
+        String profileImageUrl
+) {}
