+++ src/main/java/com/ktb/cafeboo/domain/coffeechat/dto/common/WriterDto.java
@@ -1,0 +1,6 @@
+package com.ktb.cafeboo.domain.coffeechat.dto.common;
+
+public record WriterDto(
+        String nickname,
+        String profileImageUrl
+) {}
